synthetic-telemetry-diffusion/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "synthetic-telemetry"
version = "0.1.0"
description = "Offline vector diffusion for synthetic MasterMind / Ember telemetry windows"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "torch>=2.1",
  "PyYAML>=6",
  "pandas>=2",
  "tqdm>=4.65",
]

[tool.setuptools.packages.find]
where = ["src"]
